requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiomysql==0.2.0
aiosignal==1.4.0
aiosqlite==0.21.0
anyio==4.10.0
async-timeout==5.0.1
attrs==23.2.0
Automat==22.10.0
blinker==1.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
constantly==23.10.4
Cython==3.0.8
discord-webhook==1.3.1
discord.py==2.5.2
exceptiongroup==1.3.0
Flask==3.1.1
flask-cors==6.0.1
Flask-MySQLdb==2.0.0
frozenlist==1.7.0
geoip2==5.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsons==1.6.3
lupa==2.0
MarkupSafe==3.0.2
maxminddb==2.8.2
multidict==6.6.3
mysql-connector-python==9.4.0
mysqlclient==2.2.7
pefile==2024.8.26
pillow==10.2.0
propcache==0.3.2
PyMySQL==1.1.0
python-dotenv==1.1.1
python-telegram-bot==22.3
requests==2.31.0
six==1.16.0
sniffio==1.3.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.9.0
typish==1.9.3
urllib3==2.2.0
Werkzeug==3.1.3
yarl==1.20.1
zope.interface==6.1
